+++ .gitignore
@@ -24,5 +24,6 @@ _testmain.go
 *.prof
 
 # cfs binary
-server
-ctl
+/server/server
+/server/cfs0000
+/ctl/ctl
